segmentation_processing/requirements-lock.txt
absl-py==1.2.0
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
applicationinsights==0.11.10
argcomplete==2.0.0
astroid==2.12.10
async-timeout==4.0.2
attrs==22.1.0
azure-common==1.1.28
azure-core==1.25.1
azure-graphrbac==0.61.1
azure-identity==1.7.0
azure-mgmt-authorization==2.0.0
azure-mgmt-containerregistry==10.0.0
azure-mgmt-core==1.3.2
azure-mgmt-keyvault==10.1.0
azure-mgmt-resource==21.1.0
azure-mgmt-storage==20.0.0
azureml-automl-core==1.45.0
azureml-core==1.45.0.post1
azureml-dataprep==4.2.2
azureml-dataprep-native==38.0.0
azureml-dataprep-rslex==2.8.1
azureml-dataset-runtime==1.45.0
azureml-pipeline==1.45.0
azureml-pipeline-core==1.45.0
azureml-pipeline-steps==1.45.0
azureml-sdk==1.45.0
azureml-telemetry==1.45.0
azureml-train-automl-client==1.45.0
azureml-train-core==1.45.0
azureml-train-restclients-hyperdrive==1.45.0
Babel==2.10.3
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.0.0
black==22.8.0
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-default-group==1.2.2
cloudpickle==2.2.0
contextlib2==21.6.0
contourpy==1.0.5
coverage==6.4.4
cryptography==37.0.4
cycler==0.11.0
dill==0.3.5.1
distro==1.7.0
docker==5.0.3
docutils==0.17.1
dotnetcore2==3.1.23
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.37.3
frozenlist==1.3.1
fsspec==2022.8.2
fusepy==3.0.1
google-auth==2.11.1
google-auth-oauthlib==0.4.6
grpcio==1.48.1
gunicorn==20.1.0
humanfriendly==10.0
idna==3.4
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==21.3.0
iniconfig==1.1.1
isodate==0.6.1
isort==5.10.1
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpickle==2.2.0
jsonschema==4.16.0
kiwisolver==1.4.4
knack==0.9.0
lazy-object-proxy==1.7.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
mccabe==0.7.0
msal==1.18.0
msal-extensions==0.3.1
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.2
mypy==0.971
mypy-extensions==0.4.3
ndg-httpsclient==0.5.1
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
paramiko==2.11.0
pathspec==0.10.1
Pillow==9.2.0
pipdeptree==2.3.1
pkginfo==1.8.3
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
portalocker==2.5.1
protobuf==3.19.5
py==1.11.0
py-cpuinfo==8.0.0
pyarrow==6.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pyDeprecate==0.3.2
pyflakes==2.5.0
Pygments==2.13.0
PyJWT==2.5.0
pylint==2.15.3
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
pytest-benchmark==3.4.1
pytest-cov==3.0.0
pytest-pythonpath==0.7.4
python-dateutil==2.8.2
pytorch-lightning==1.7.6
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.1
seaborn==0.12.0
SecretStorage==3.3.3
six==1.16.0
snowballstemmer==2.2.0
Sphinx==5.1.1
sphinx-autodoc-typehints==1.19.2
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch==1.12.1
torch-metrics==1.1.7
torchmetrics==0.9.3
torchvision==0.13.1
towncrier==22.8.0
tqdm==4.64.1
types-cryptography==3.3.23
types-PyYAML==6.0.11
types-requests==2.28.10
types-tabulate==0.8.11
types-urllib3==1.26.24
typing_extensions==4.3.0
urllib3==1.26.12
websocket-client==1.4.1
Werkzeug==2.2.2
wget==3.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
